[refactor] Cacher => CollectionEnumerator
