+++ lib/movienga/collection_enumerator.rb
@@ -1,0 +1,19 @@
+require 'ruby-progressbar'
+
+module Movienga
+  class CollectionEnumerator
+    using ProgressBar::Refinements::Enumerator
+
+    def initialize(collection)
+      @collection = collection
+    end
+
+    def run(title: 'Caching', &block)
+      @collection.each.with_progressbar(
+        title: title,
+        format: "%t: Processed %c/%C |%b>%i| %p%% %e %a",
+        &block
+      )
+    end
+  end
+end
